config-extractor/requirements.txt
requests==2.26.0
PyYAML==5.4.1
tqdm==4.62.3